backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "vivaldi",email = "hross8888@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "django-ninja (>=1.4.1,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "redis (>=6.0.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
	"gunicorn (>=23.0.0,<24.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-cors-headers = "^4.7.0"
mypy = "^1.15.0"
bandit = "^1.8.3"
ruff = "^0.11.9"

[tool.mypy]
ignore_missing_imports = true


[tool.ruff]
line-length = 120
target-version = "py311"
lint.select = ["E", "F", "I"]
exclude = [
    ".venv",
    "*/migrations/*"
]